backend/requirements.txt
Django==3.2.10
djangorestframework==3.12.4
Pillow==8.4.0
torch==1.10.0
torchvision==0.11.1
timm